Safari-609.1.11.1/LayoutTests/editing/spelling/spelling-backspace-between-lines-expected.txt

Spelling markers should remain while merging two lines.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS firstLineText('target1') is "OK"
PASS sel.anchorNode.data is "OK zz OK"
PASS internals.hasSpellingMarker(3, 2) became true
PASS firstLineText('target2') is "OKOK zz OK"
PASS sel.anchorNode.data is "OKOK zz OK"
PASS internals.hasSpellingMarker(5, 2) became true
PASS firstLineText('target3') is "OKOK zz OK"
PASS sel.anchorNode.data is "OKOK zz OK"
PASS internals.hasSpellingMarker(5, 2) became true
PASS successfullyParsed is true

TEST COMPLETE
